[url] Properly treat '#' as marking the start of the fragment in data URIs
This CL aligns Chromium with the URL spec, such that we consider '#' to
mark the end of the content and the start of the fragment section only.
GURL::GetContent was updated to reflect this, with a special case for
javascript URLs specifically (as their spec mentions including '#' in the
URL content).

Bug: 123004
Cq-Include-Trybots: master.tryserver.chromium.linux:closure_compilation;master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I49c6f4d739a2dd42eecc9947f8e75071b84e9be7